Build other office class
