+++ spec/remote_offices_spec.rb
@@ -1,0 +1,5 @@
+require 'spec_helper'
+
+RSpec.describe RemoteOffices do
+  describe
+end
